Updated embark_screen with new map updates
We're no longer generating separate world and regional maps, instead
there is only 1 map now, the whole world.

The "world map" is now just a display of the MapCells of the single map,
and the "regional map" is the MapLocations of a MapCell.

For generational purposes, the world map display is 1:1 ratio, i.e. each
place along on the x axis is +1, and likewise for y. For the regional
map, each place along is 1/map_cell_width along. This allows us to "zoom
in" the noise function